IOS-304 Fix CI unit tests build + remove Xcode proj
Ideally we'd be using `swift test` to run the unit tests for the Swift Package, but that fails because it always seem to want to build for macOS platform, and then erroring out because UIKit is not available. This is despite passing `-Xswiftc "-target" -Xswiftc "x86_64-apple-ios13.0-simulator"` to `swift test|build`.

Also removing the Xcode project entirely because (1) xcodebuild does not allow to select a non-project file such as Package.swift and instead defaults to the Xcode project if present, and (2) the option to generate an Xcode project from the Swift Package is deprecated.